Closes #17
