+++ src/test/java/im/status/wallet/TestSecureChannelSession.java
@@ -1,0 +1,9 @@
+package im.status.wallet;
+
+import im.status.hardwallet.lite.SecureChannelSession;
+
+public class TestSecureChannelSession extends SecureChannelSession {
+  public void setOpen() {
+    super.setOpen();
+  }
+}
